Add Sidebar component to Admin and Accounts pages
